requirements.txt
torch>=1.10.0
torchvision>=0.11.0
matplotlib>=3.4.0
numpy>=1.20.0
pillow>=8.0.0